Use Avo

+++ Example/IosAnalyticsDebugger/Avo.m
@@ -237,7 +237,7 @@ static NSObject * __DEBUGGER__ = nil;
   return @[];
 }
 
-+ (NSArray *)assertNonNullEnum:(NSString *)propertyId propertyName:(NSString *)propertyName value:(int)value
++ (NSArray *)assertNonNullEnum:(NSString *)propertyId propertyName:(NSString *)propertyName value:(unsigned long)value
 {
   if (value < 0) {
     return @[@{@"tag": @"expectedInitializedEnum", @"propertyId": propertyId,
